// src/enums/playreq.ts
export const enum PlayReq {
  INVALID = -1,
  REQ_MINION_TARGET = 1,
  REQ_FRIENDLY_TARGET = 2,
  REQ_ENEMY_TARGET = 3,
  REQ_DAMAGED_TARGET = 4,
  REQ_FREEZE_TARGET = 5,
  REQ_CHARGE_TARGET = 6,
  REQ_TARGET_MAX_ATTACK = 7,
  REQ_UNDAMAGED_TARGET = 8,
  REQ_HERO_TARGET = 9,
  REQ_SECRET_ZONE_CAP = 10,
  REQ_SECRET_CAP = 11,
  REQ_SECRET_CAP_3 = 12,
  REQ_ATTACK_GREATER_THAN_0 = 13,
  REQ_ATTACK_BETWEEN_ENEMY_MINIONS = 14,
  REQ_MINION_CAP = 15,
  REQ_MINION_CAP_3 = 16,
  REQ_TARGET_IF_AVAILABLE = 17,
  REQ_MINIMUM_ENEMY_MINIONS = 18,
  REQ_TARGET_FOR_COMBO = 19,
  REQ_NOT_MINION_DESTROYED = 20,
  REQ_NOT_EXHAUSTED_SOURCE = 21,
  REQ_NOT_EXHAUSTED_TARGET = 22,
  REQ_NOT_SECRET_TARGET = 23,
  REQ_STEALTHED_TARGET = 24,
  REQ_TARGET_NOT_STEALTHED = 25,
  REQ_ATTACKABLE_BY_COMBO = 26,
  REQ_TARGET_TAUNT = 27,
  REQ_TARGET_SPELLPOWER = 28,
  REQ_TARGET_RACE = 29,
  REQ_TARGET_IS_NOT_SELF = 30,
  REQ_NON_SECRET_TARGET = 31,
  REQ_TARGET_IS_MAX_ATTACK = 32,
  REQ_TARGET_NOT_DESTROYED = 33,
  REQ_NOT_HAND_TARGET = 34,
  REQ_TARGET_NOT_GRAVEYARD = 35,
  REQ_TARGET_NOT_LOOT = 36,
  REQ_ENEMY_TARGET_NOT_LOOT = 37,
  REQ_TARGET_NOT_MARKED = 38,
  REQ_BATTLECRIE_QUALIFIER = 39,
  REQ_TARGET_6_COST = 40,
  REQ_TARGET_7_COST = 41,
  REQ_TARGET_8_COST = 42,
  REQ_TARGET_9_COST = 43,
  REQ_TARGET_10_COST = 44,
  REQ_TARGET_MIN_ATTACK = 45,
  REQ_CAN_BE_ATTACKED = 46,
  REQ_TARGET_SLUDGE_COST = 47,
  REQ_TARGET_TO_PLAY = 48,
  REQ_TARGET_WITH_DEATHRATTLE = 49,
  REQ_TARGET_WOTOG = 50,
  REQ_TARGET_IS_RACE = 51,
  REQ_TARGET_NOT_DORMANT = 52,
  REQ_TARGET_IS_GUARDIAN = 53,
  REQ_TARGET_IS_NOT_ASLEEP = 54,
  REQ_TARGET_NOT_BURNED = 55,
  REQ_NUM_MINION_SLOTS = 56,
  REQ_WEAPON_EQUIPPED = 57,
}